+++ src/Aardvark.Cef.WinForms/Browser.fs
@@ -4,10 +4,16 @@ open CefSharp
 open CefSharp.Web
 open CefSharp.WinForms
 open System
-open System.Runtime.InteropServices
 
 type AardvarkCefBrowser =
     inherit ChromiumWebBrowser
+
+    static let disableZoomJs = @"
+        window.addEventListener('wheel', function(e) {
+            if (e.ctrlKey) {
+                e.preventDefault();
+            }
+        }, { passive: false });"
 
     private new (address: string, requestContext: IRequestContext) =
         { inherit ChromiumWebBrowser(address, requestContext) }
@@ -15,14 +21,22 @@ type AardvarkCefBrowser =
     private new (html: HtmlString, requestContext: IRequestContext) =
         { inherit ChromiumWebBrowser(html, requestContext) }
 
+    member private this.Initialize() =
+        AardvarkRenderProcessMessageHandler.Setup this
+
+        this.FrameLoadEnd.Add (fun args ->
+            if args.Frame.IsMain && args.Frame.IsValid then
+                args.Frame.ExecuteJavaScriptAsync(disableZoomJs)
+        )
+
     static member internal Create(address: string, requestContext: IRequestContext) =
         let browser = new AardvarkCefBrowser(address, requestContext)
-        AardvarkRenderProcessMessageHandler.Setup browser
+        browser.Initialize()
         browser
 
     static member internal Create(html: HtmlString, requestContext: IRequestContext) =
         let browser = new AardvarkCefBrowser(html, requestContext)
-        AardvarkRenderProcessMessageHandler.Setup browser
+        browser.Initialize()
         browser
 
     member this.OpenDevTools() =